requirements.txt
PyYAML~=6.0.2
paho-mqtt~=2.1.0
python-dotenv~=1.0.1
Flask~=3.1.0
Hypercorn~=0.17.3
discord-webhook~=1.3.1
ring-doorbell~=0.9.12